requirements.txt
Flask==2.3.2
Flask-Mail==0.9.1
google-generativeai==0.3.0
Werkzeug==2.3.7
